Remove RandomTmpDirPath
